Take into use a new capability to define project per device.
This requires using Cordova scripts that has implementation for this
https://issues.apache.org/jira/browse/CB-8123 feature.

When using older scripts, builds for Windows and Windows Phone does
not work out of box. With older scripts, one must comment out one
of the referenced projects and then builds will work only to the target
that remains in plugin.xml file.