Start implementing a common API across all platforms.
The API is defined in www/AllJoyn.js and it is expected that
platform-specific implementations of the plugin implement the
cordova.exec calls made from that file.

After this change, the API is accessible as "AllJoyn" global object.

In the Windows version of the plugin, there is a JavaScript layer
that handles the cordova.exec calls in file src/windows/AllJoynProxy.js.

This is just an initial structure for the API and not all functionality is
yet implemented and tested.